Rename main to action
